UPDATE: replace interface with implementation
