+++ src/test/java/mops/businesslogic/FileQueryTest.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class FileQueryImplTest {
+public class FileQueryTest {
 
     private FileInfo fileInfo;
